Factor out CreateRoleTask from TransportCreateRoleAction
